MueLu: Move simple driver to examples
