+++ packages/muelu/example/basic/Simple.cpp
@@ -43,106 +43,94 @@
 // ***********************************************************************
 //
 // @HEADER
-//
-// To compile and run this example, Trilinos must be configured with
-// Tpetra, Amesos2, MueLu, Ifpack2, and Belos.
-//
-// This example will only work with Tpetra, not Epetra.
-//
-// Commonly used options are
-//   --matrixType
-//   --nx
-//   --ny
-//   --xmlFile
-//
-// "./MueLu_Simple.exe --help" shows all supported command line options.
-//
-
+#include <cstdio>
+#include <iomanip>
 #include <iostream>
+#include <unistd.h>
 
-// Tpetra provides distributed sparse linear algebra
-#include <Tpetra_CrsMatrix.hpp>
-#include <Tpetra_Vector.hpp>
+#include <Teuchos_XMLParameterListHelpers.hpp>
+#include <Teuchos_YamlParameterListHelpers.hpp>
+#include <Teuchos_StandardCatchMacros.hpp>
 
-// Belos provides Krylov solvers
-#include <BelosConfigDefs.hpp>
-#include <BelosLinearProblem.hpp>
-#include <BelosBlockCGSolMgr.hpp>
-#include <BelosPseudoBlockCGSolMgr.hpp>
-#include <BelosBlockGmresSolMgr.hpp>
-#include <BelosTpetraAdapter.hpp>
+// Xpetra
+#include <Xpetra_MultiVectorFactory.hpp>
+#include <Xpetra_Operator.hpp>
+#include <Xpetra_Map.hpp>
+#include <Xpetra_MultiVector.hpp>
 
 // Galeri
 #include <Galeri_XpetraParameters.hpp>
 #include <Galeri_XpetraProblemFactory.hpp>
+#include <Galeri_XpetraUtils.hpp>
+#include <Galeri_XpetraMaps.hpp>
 
-// MueLu main header: include most common header files in one line
 #include <MueLu.hpp>
-#include <MueLu_TpetraOperator.hpp>
-#include <MueLu_CreateTpetraPreconditioner.hpp>
+
+#include <MueLu_BaseClass.hpp>
+#ifdef HAVE_MUELU_EXPLICIT_INSTANTIATION
+#include <MueLu_ExplicitInstantiation.hpp>
+#endif
+#include <MueLu_Level.hpp>
+#include <MueLu_MutuallyExclusiveTime.hpp>
+#include <MueLu_ParameterListInterpreter.hpp>
 #include <MueLu_Utilities.hpp>
+#include <MatrixLoad.hpp>
+#include <DriverCore.hpp>
 
-int main(int argc, char *argv[]) {
 
-  // Define default types
-  typedef double                                      scalar_type;
-  typedef int                                         local_ordinal_type;
-  typedef int                                         global_ordinal_type;
-  typedef KokkosClassic::DefaultNode::DefaultNodeType node_type;
+/*********************************************************************/
 
-  // Convenient typedef's
-  typedef Tpetra::Operator<scalar_type,local_ordinal_type,global_ordinal_type,node_type>    operator_type;
-  typedef Tpetra::CrsMatrix<scalar_type,local_ordinal_type,global_ordinal_type,node_type>   crs_matrix_type;
-  typedef Tpetra::Vector<scalar_type,local_ordinal_type,global_ordinal_type,node_type>      vector_type;
-  typedef Tpetra::MultiVector<scalar_type,local_ordinal_type,global_ordinal_type,node_type> multivector_type;
-  typedef Tpetra::Map<local_ordinal_type,global_ordinal_type,node_type>                     driver_map_type;
+template<class Scalar, class LocalOrdinal, class GlobalOrdinal, class Node>
+int main_(Teuchos::CommandLineProcessor &clp, Xpetra::UnderlyingLib& lib, int argc, char *argv[]) {
+#include <MueLu_UseShortNames.hpp>
+  using Teuchos::RCP;
+  using Teuchos::rcp;
+  using Teuchos::ArrayRCP;
+  using Teuchos::TimeMonitor;
+  using Teuchos::ParameterList;
 
-  typedef MueLu::TpetraOperator<scalar_type, local_ordinal_type, global_ordinal_type, node_type> muelu_tpetra_operator_type;
-  typedef MueLu::Utilities<scalar_type,local_ordinal_type,global_ordinal_type,node_type> MueLuUtilities;
-
-  typedef Belos::LinearProblem<scalar_type, multivector_type, operator_type> linear_problem_type;
-  typedef Belos::SolverManager<scalar_type, multivector_type, operator_type> belos_solver_manager_type;
-  typedef Belos::PseudoBlockCGSolMgr<scalar_type, multivector_type, operator_type> belos_pseudocg_manager_type;
-  typedef Belos::BlockGmresSolMgr<scalar_type, multivector_type, operator_type> belos_gmres_manager_type;
-
-  //MueLu_UseShortNames.hpp wants these typedefs.
-  typedef scalar_type         Scalar;
-  typedef local_ordinal_type  LocalOrdinal;
-  typedef global_ordinal_type GlobalOrdinal;
-  typedef node_type           Node;
-# include <MueLu_UseShortNames.hpp>
-
-  typedef Galeri::Xpetra::Problem<Map,CrsMatrixWrap,MultiVector> GaleriXpetraProblem;
-
-  using Teuchos::RCP; // reference count pointers
-  using Teuchos::rcp; // reference count pointers
-
-  //
+  // =========================================================================
   // MPI initialization using Teuchos
-  //
+  // =========================================================================
+  RCP<const Teuchos::Comm<int> > comm = Teuchos::DefaultComm<int>::getComm();
 
-  Teuchos::GlobalMPISession mpiSession(&argc, &argv, NULL);
-  RCP< const Teuchos::Comm<int> > comm = Teuchos::DefaultComm<int>::getComm();
-  int mypid = comm->getRank();
+  // =========================================================================
+  // Convenient definitions
+  // =========================================================================
+  typedef Teuchos::ScalarTraits<SC> STS;
+  typedef typename STS::coordinateType real_type;
+  typedef Xpetra::MultiVector<real_type,LO,GO,NO> RealValuedMultiVector;
 
-  Teuchos::CommandLineProcessor clp(false);
+  // =========================================================================
+  // Parameters initialization
+  // =========================================================================
+  GO nx = 100, ny = 100, nz = 100;
+  Galeri::Xpetra::Parameters<GO> galeriParameters(clp, nx, ny, nz, "Laplace2D"); // manage parameters of the test case
+  Xpetra::Parameters             xpetraParameters(clp);                          // manage parameters of Xpetra
 
-  // Parameters
+  std::string xmlFileName       = "";                 clp.setOption("xml",                   &xmlFileName,       "read parameters from an xml file");
+  std::string yamlFileName      = "";                 clp.setOption("yaml",                  &yamlFileName,      "read parameters from a yaml file");
+  bool        printTimings      = true;               clp.setOption("timings", "notimings",  &printTimings,      "print timings to screen");
+  std::string timingsFormat     = "table-fixed";      clp.setOption("time-format",           &timingsFormat,     "timings format (table-fixed | table-scientific | yaml)");
+  std::string solveType         = "belos";            clp.setOption("solver",                &solveType,         "solve type: (none | belos)");
+  std::string belosType         = "cg";               clp.setOption("belosType",             &belosType,         "belos solver type: (Pseudoblock CG | Block CG | Pseudoblock GMRES | Block GMRES | ...) see BelosSolverFactory.hpp for exhaustive list of solvers");
+  double      tol               = 1e-12;              clp.setOption("tol",                   &tol,              "solver convergence tolerance");
+  bool        binaryFormat      = false;              clp.setOption("binary", "ascii",       &binaryFormat,      "read matrices in binary format");
+  std::string rowMapFile;                             clp.setOption("rowmap",                &rowMapFile,        "map data file");
+  std::string colMapFile;                             clp.setOption("colmap",                &colMapFile,        "colmap data file");
+  std::string domainMapFile;                          clp.setOption("domainmap",             &domainMapFile,     "domainmap data file");
+  std::string rangeMapFile;                           clp.setOption("rangemap",              &rangeMapFile,      "rangemap data file");
+  std::string matrixFile;                             clp.setOption("matrix",                &matrixFile,        "matrix data file");
+  std::string rhsFile;                                clp.setOption("rhs",                   &rhsFile,           "rhs data file");
+  std::string coordFile;                              clp.setOption("coords",                &coordFile,         "coordinates data file");
+  std::string nullFile;                               clp.setOption("nullspace",             &nullFile,          "nullspace data file");
+  int         maxIts            = 200;                clp.setOption("its",                   &maxIts,            "maximum number of solver iterations");
+  int         numVectors        = 1;                  clp.setOption("multivector",           &numVectors,        "number of rhs to solve simultaneously");
+  bool        scaleResidualHist = true;               clp.setOption("scale", "noscale",      &scaleResidualHist, "scaled Krylov residual history");
+  bool        solvePreconditioned = true;             clp.setOption("solve-preconditioned","no-solve-preconditioned", &solvePreconditioned, "use MueLu preconditioner in solve");
+  int         cacheSize = 0;                          clp.setOption("cachesize",             &cacheSize,         "cache size (in KB)");
 
-  global_ordinal_type nx = 50;
-  Galeri::Xpetra::Parameters<GO> matrixParameters(clp, nx); // manage parameters of the test case
-  Xpetra::Parameters             xpetraParameters(clp);     // manage parameters of xpetra
-
-  global_ordinal_type maxIts            = 25;
-  scalar_type tol                       = 1e-8;
-  std::string solverOptionsFile         = "amg.xml";
-  std::string krylovSolverType          = "cg";
-
-  clp.setOption("xmlFile",    &solverOptionsFile, "XML file containing MueLu solver parameters");
-  clp.setOption("maxits",     &maxIts,            "maximum number of Krylov iterations");
-  clp.setOption("tol",        &tol,               "tolerance for Krylov solver");
-  clp.setOption("krylovType", &krylovSolverType,  "cg or gmres solver");
-
+  clp.recogniseAllOptions(true);
   switch (clp.parse(argc, argv)) {
     case Teuchos::CommandLineProcessor::PARSE_HELP_PRINTED:        return EXIT_SUCCESS;
     case Teuchos::CommandLineProcessor::PARSE_ERROR:
@@ -150,76 +138,118 @@ int main(int argc, char *argv[]) {
     case Teuchos::CommandLineProcessor::PARSE_SUCCESSFUL:          break;
   }
 
-  if (xpetraParameters.GetLib() == Xpetra::UseEpetra) {
-    throw std::invalid_argument("This example only supports Tpetra.");
+  TEUCHOS_TEST_FOR_EXCEPTION(xmlFileName != "" && yamlFileName != "", std::runtime_error,
+                             "Cannot provide both xml and yaml input files");
+
+  // Instead of checking each time for rank, create a rank 0 stream
+  RCP<Teuchos::FancyOStream> fancy = Teuchos::fancyOStream(Teuchos::rcpFromRef(std::cout));
+  Teuchos::FancyOStream& out = *fancy;
+  out.setOutputToRootOnly(0);
+
+  ParameterList paramList;
+  auto inst = xpetraParameters.GetInstantiation();
+
+  if (yamlFileName != "") {
+    Teuchos::updateParametersFromYamlFileAndBroadcast(yamlFileName, Teuchos::Ptr<ParameterList>(&paramList), *comm);
+  } else {
+    xmlFileName = (xmlFileName != "" ? xmlFileName : "simple.xml");
+    Teuchos::updateParametersFromXmlFileAndBroadcast(xmlFileName, Teuchos::Ptr<ParameterList>(&paramList), *comm);
   }
 
-  Teuchos::ParameterList mueluParams;
-  Teuchos::updateParametersFromXmlFile(solverOptionsFile, Teuchos::inoutArg(mueluParams));
-
-  //
-  // Construct the problem
-  //
-
-  global_ordinal_type indexBase = 0;
-  RCP<const Map> xpetraMap = MapFactory::Build(Xpetra::UseTpetra, matrixParameters.GetNumGlobalElements(), indexBase, comm);
-  RCP<GaleriXpetraProblem> Pr = Galeri::Xpetra::BuildProblem<scalar_type, local_ordinal_type, global_ordinal_type, Map, CrsMatrixWrap, MultiVector>(matrixParameters.GetMatrixType(), xpetraMap, matrixParameters.GetParameterList());
-  RCP<Matrix>  xpetraA = Pr->BuildMatrix();
-  RCP<crs_matrix_type> A = MueLuUtilities::Op2NonConstTpetraCrs(xpetraA);
-  RCP<const driver_map_type> map = MueLuUtilities::Map2TpetraMap(*xpetraMap);
-
-  //
-  // Construct a multigrid preconditioner
-  //
-
-  // Multigrid Hierarchy
-  RCP<muelu_tpetra_operator_type> M = MueLu::CreateTpetraPreconditioner((RCP<operator_type>)A, mueluParams);
-
-  //
-  // Set up linear problem Ax = b and associate preconditioner with it.
-  //
-
-  RCP<multivector_type> X = rcp(new multivector_type(map,1));
-  RCP<multivector_type> B = rcp(new multivector_type(map,1));
-
-  X->putScalar((scalar_type) 0.0);
-  B->randomize();
-
-  RCP<linear_problem_type> Problem = rcp(new linear_problem_type(A, X, B));
-  Problem->setRightPrec(M);
-  Problem->setProblem();
-
-  //
-  // Set up Krylov solver and iterate.
-  //
-  RCP<Teuchos::ParameterList> belosList = rcp(new Teuchos::ParameterList());
-  belosList->set("Maximum Iterations",    maxIts); // Maximum number of iterations allowed
-  belosList->set("Convergence Tolerance", tol);    // Relative convergence tolerance requested
-  belosList->set("Verbosity",             Belos::Errors + Belos::Warnings + Belos::StatusTestDetails);
-  belosList->set("Output Frequency",      1);
-  belosList->set("Output Style",          Belos::Brief);
-  belosList->set("Implicit Residual Scaling", "None");
-  RCP<belos_solver_manager_type> solver;
-  if (krylovSolverType == "cg")
-    solver = rcp(new belos_pseudocg_manager_type(Problem, belosList));
-  else if (krylovSolverType == "gmres")
-    solver = rcp(new belos_gmres_manager_type(Problem, belosList));
-  else
-    throw std::invalid_argument("bad Krylov solver type");
-
-  solver->solve();
-  int numIterations = solver->getNumIters();
-
-  Teuchos::Array<typename Teuchos::ScalarTraits<scalar_type>::magnitudeType> normVec(1);
-  multivector_type Ax(B->getMap(),1);
-  multivector_type residual(B->getMap(),1);
-  A->apply(*X, residual);
-  residual.update(1.0, *B, -1.0);
-  residual.norm2(normVec);
-  if (mypid == 0) {
-    std::cout << "number of iterations = " << numIterations << std::endl;
-    std::cout << "||Residual|| = " << normVec[0] << std::endl;
+  if (inst == Xpetra::COMPLEX_INT_INT && solveType == "belos") {
+    belosType = "gmres";
+    out << "WARNING: CG will not work with COMPLEX scalars, switching to GMRES"<<std::endl;
   }
+
+  // Retrieve matrix parameters (they may have been changed on the command line)
+  // [for instance, if we changed matrix type from 2D to 3D we need to update nz]
+  ParameterList galeriList = galeriParameters.GetParameterList();
+
+  // =========================================================================
+  // Problem construction
+  // =========================================================================
+  std::ostringstream galeriStream;
+
+  comm->barrier();
+  Teuchos::TimeMonitor::setStackedTimer(Teuchos::null);
+  RCP<TimeMonitor> globalTimeMonitor = rcp(new TimeMonitor(*TimeMonitor::getNewTimer("Driver: S - Global Time")));
+  RCP<TimeMonitor> tm                = rcp(new TimeMonitor(*TimeMonitor::getNewTimer("Driver: 1 - Matrix Build")));
+
+  RCP<Matrix>                 A;
+  RCP<const Map>              map;
+  RCP<RealValuedMultiVector> coordinates;
+  RCP<MultiVector> nullspace;
+  RCP<MultiVector> X, B;
+
+  // Load the matrix off disk (or generate it via Galeri)
+  MatrixLoad<SC,LO,GO,NO>(comm,lib,binaryFormat,matrixFile,rhsFile,rowMapFile,colMapFile,domainMapFile,rangeMapFile,coordFile,nullFile,map,A,coordinates,nullspace,X,B,numVectors,galeriParameters,xpetraParameters,galeriStream);
+  comm->barrier();
+  tm = Teuchos::null;
+  out << galeriStream.str();
+
+  // =========================================================================
+  // Preconditioner construction
+  // =========================================================================
+  bool useML   = paramList.isParameter("use external multigrid package") && (paramList.get<std::string>("use external multigrid package") == "ml");
+  out<<"*********** MueLu ParameterList ***********"<<std::endl;
+  out<<paramList;
+  out<<"*******************************************"<<std::endl;
+
+  RCP<Hierarchy> H;
+  RCP<Operator> Prec;
+  {
+    comm->barrier();
+    tm = rcp(new TimeMonitor(*TimeMonitor::getNewTimer("Driver: 2 - MueLu Setup")));
+    PreconditionerSetup(A,coordinates,nullspace,paramList,false,false,useML,0,H,Prec);
+    comm->barrier();
+    tm = Teuchos::null;
+  }
+
+  // =========================================================================
+  // System solution (Ax = b)
+  // =========================================================================
+  {
+    comm->barrier();
+    SystemSolve(A,X,B,H,Prec,out,solveType,belosType,false,false,useML,cacheSize,0,scaleResidualHist,solvePreconditioned,maxIts,tol);
+    comm->barrier();
+  }
+
+  tm = Teuchos::null;
+  globalTimeMonitor = Teuchos::null;
+
+  if (printTimings) {
+    RCP<ParameterList> reportParams = rcp(new ParameterList);
+    if (timingsFormat == "yaml") {
+      reportParams->set("Report format",             "YAML");            // "Table" or "YAML"
+      reportParams->set("YAML style",                "compact");         // "spacious" or "compact"
+    }
+    reportParams->set("How to merge timer sets",   "Union");
+    reportParams->set("alwaysWriteLocal",          false);
+    reportParams->set("writeGlobalStats",          true);
+    reportParams->set("writeZeroTimers",           false);
+
+    const std::string filter = "";
+
+    std::ios_base::fmtflags ff(out.flags());
+    if (timingsFormat == "table-fixed") out << std::fixed;
+    else                                out << std::scientific;
+    TimeMonitor::report(comm.ptr(), out, filter, reportParams);
+    out << std::setiosflags(ff);
+  }
+
+  TimeMonitor::clearCounters();
+  out << std::endl;
+
+  fflush(NULL);
+  comm->barrier();
 
   return EXIT_SUCCESS;
+}
+
+//- -- --------------------------------------------------------
+#define MUELU_AUTOMATIC_TEST_ETI_NAME main_
+#include "MueLu_Test_ETI.hpp"
+
+int main(int argc, char *argv[]) {
+  return Automatic_Test_ETI(argc,argv);
 }
