--- packages/muelu/example/basic/amg.xml
@@ -1,32 +1,0 @@
-<ParameterList name="MueLu">
-  <Parameter        name="verbosity"                            type="string"   value="high"/>
-
-  <Parameter        name="number of equations"                  type="int"      value="1"/>
-
-  <Parameter        name="coarse: max size"                     type="int"      value="500"/>
-
-  <Parameter        name="multigrid algorithm"                  type="string"   value="sa"/>
-
-  <!-- ===========  AGGREGATION  =========== -->
-  <Parameter        name="aggregation: type"                    type="string"   value="uncoupled"/>
-  <Parameter        name="aggregation: drop scheme"             type="string"   value="classical"/>
-  <!-- <Parameter        name="aggregation: drop tol"                type="double"   value="0.1"/> -->
-
-  <!-- ===========  SMOOTHING  =========== -->
-  <Parameter        name="smoother: type"                       type="string"   value="CHEBYSHEV"/>
-  <ParameterList    name="smoother: params">
-    <Parameter      name="chebyshev: degree"                    type="int"      value="2"/>>
-    <Parameter      name="chebyshev: ratio eigenvalue"          type="double"   value="7"/>
-    <Parameter      name="chebyshev: min eigenvalue"            type="double"   value="1.0"/>
-    <Parameter      name="chebyshev: zero starting solution"    type="bool"     value="true"/>
-  </ParameterList>
-
-  <!-- ===========  REPARTITIONING  =========== -->
-  <Parameter        name="repartition: enable"                  type="bool"     value="false"/>
-  <Parameter        name="repartition: partitioner"             type="string"   value="zoltan"/>
-  <Parameter        name="repartition: start level"             type="int"      value="2"/>
-  <Parameter        name="repartition: min rows per proc"       type="int"      value="800"/>
-  <Parameter        name="repartition: max imbalance"           type="double"   value="1.1"/>
-  <Parameter        name="repartition: remap parts"             type="bool"     value="false"/>
-
-</ParameterList>
